Separate Lambda and AWS OTel features
